Rename 'notifications' stuff to 'post' stuff
