+++ app/src/main/java/com/example/cosc341_project/ui/post/PostFragment.java
@@ -1,0 +1,37 @@
+package com.example.cosc341_project.ui.post;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.cosc341_project.databinding.FragmentPostBinding;
+
+public class PostFragment extends Fragment {
+
+    private FragmentPostBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        PostViewModel postViewModel =
+                new ViewModelProvider(this).get(PostViewModel.class);
+
+        binding = FragmentPostBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+
+        final TextView textView = binding.textPost;
+        postViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
