+++ app/src/main/java/com/example/cosc341_project/ui/post/PostViewModel.java
@@ -1,16 +1,16 @@
-package com.example.cosc341_project.ui.notifications;
+package com.example.cosc341_project.ui.post;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class PostViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public PostViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is post fragment");
     }
 
     public LiveData<String> getText() {
