+++ app/src/main/res/layout/fragment_post.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.NotificationsFragment">
+    tools:context=".ui.post.PostFragment">
 
     <TextView
-        android:id="@+id/text_notifications"
+        android:id="@+id/text_post"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
