+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,14 +7,14 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
-
-    <item
         android:id="@id/navigation_map"
         android:icon="@drawable/baseline_map_24"
         android:title="@string/title_map" />
+
+    <item
+        android:id="@+id/navigation_post"
+        android:icon="@drawable/baseline_add_circle_outline_24"
+        android:title="@string/title_post" />
 
     <item
         android:id="@id/navigation_profile"
