+++ app/src/main/res/values/strings.xml
@@ -1,12 +1,8 @@
 <resources>
     <string name="app_name">Cosc341_Project</string>
+
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
     <string name="title_map">Map</string>
+    <string name="title_post">Post</string>
     <string name="title_profile">Profile</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-
-
 </resources>
